+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   redimension filter
 - Add `BitmapFont` with a sample hand crafted font, add `bitmap_text` generator.
 - `Color#*` can now accept another `Color` to multiply channels
+- Add `Colors` filter with `color_multiply!` and alias `*`
+- `bitmap_text` accepts multiline strings and supports colorization
 
 ## [0.2.0] - 2025-07-21
 - Ruby Sixel encoder now sets pixel aspect ratio metadata to display correctly in Windows Terminal